Changes the main.rb file based on the requirements
