+++ main.rb
@@ -14,36 +14,13 @@ def menu
   puts options
 end
 
-def option_case(choice)
-  case choice
-  when 1
-    list_all_books
-  when 2
-    list_all_people
-  when 3
-    create_person
-  when 4
-    create_book
-  when 5
-    create_rental
-  when 6
-    list_rentals_by_id
-  else
-    puts 'Incorrect choice, choose a number between 1..9'
-  end
-end
-
-def prompt_user
-  menu
-
-  choice = gets.chomp.to_i
-  choice == 7 ? quit_app : option_case(choice)
-
-  prompt_user
+def quit_app
+  puts 'Thank you for using this app! Exiting...'
+  exit
 end
 
 def main
-  puts "\n Welcome to School Library App!"
+  puts "\n Welcome to School Library App! \n"
   app = App.new
   app.run
 end
